Reading of pool files from database.
Use HDBC (instead of Sqlite3.hsc) for the database.  It's much more
complete, even though it seems to not have proper blob support.  The
reading seems safe, but writing blobs has to be done by converting the
blobs into hex and placing them in the query.  Fortunately, we don't
have large amounts of blob data to store (just the hashes).