Move to explicit zlib >= 0.5.
There is an API change, to make this explicit.  Leave a comment in the
code as to how to change for the old API.

I have some suspicion that my usage of zlib is causing a C heap space
leak.  I will likely change to using the simple zlib interface without
lazy strings so that the zlib state can be allocated and deallocated
in a single IO operation.