Generalize the parser a bit.
This can now deal with extracting the first field a bit.  Really, we
need to do different things depending on what the first keyword is.
Although the keyword order is not specialized, the set is defined, and
we can easily keep or even discard the ones we don't know about.