Exception for for GHC 6.10.1.
The exception system uses a dynamic exception type which causes a
compilation error if the type is not determined (say when ignoring all
exceptions).  We're really only interested in ignoring IO exceptions
anyway, so use the IO's catch routine to ignore operations that we
want to allow to fail.  This allows the code to compile with either
ghc 6.10.1 or 6.8.